Move TextDataset class into utils
